+++ testapp/ng2/app/app.routes.ts
@@ -1,12 +1,10 @@
-import { provideRouter, RouterConfig } from '@angular/router';
+import { Routes, RouterModule } from '@angular/router';
 import { HomeComponent } from './home/home.component';
 import { AsyncComponent } from './async/async.component';
 
-export const routes: RouterConfig = [
+export const routes: Routes = [
   { path: '', component: HomeComponent },
   { path: 'async', component: AsyncComponent }
 ];
 
-export const APP_ROUTER_PROVIDERS = [
-  provideRouter(routes)
-]
+export const appRouting = RouterModule.forRoot(routes, { useHash: true });