+++ testapp/ng2/app/main.js
@@ -1,31 +1,5 @@
-System.register(['@angular/platform-browser-dynamic', '@angular/core', '@angular/common', './app.component', './app.routes'], function(exports_1, context_1) {
-    "use strict";
-    var __moduleName = context_1 && context_1.id;
-    var platform_browser_dynamic_1, core_1, common_1, app_component_1, app_routes_1;
-    return {
-        setters:[
-            function (platform_browser_dynamic_1_1) {
-                platform_browser_dynamic_1 = platform_browser_dynamic_1_1;
-            },
-            function (core_1_1) {
-                core_1 = core_1_1;
-            },
-            function (common_1_1) {
-                common_1 = common_1_1;
-            },
-            function (app_component_1_1) {
-                app_component_1 = app_component_1_1;
-            },
-            function (app_routes_1_1) {
-                app_routes_1 = app_routes_1_1;
-            }],
-        execute: function() {
-            platform_browser_dynamic_1.bootstrap(app_component_1.AppComponent, [
-                app_routes_1.APP_ROUTER_PROVIDERS,
-                core_1.provide(common_1.LocationStrategy, { useClass: common_1.HashLocationStrategy })
-            ])
-                .catch(function (err) { return console.error(err); });
-        }
-    }
-});
+"use strict";
+var platform_browser_dynamic_1 = require('@angular/platform-browser-dynamic');
+var app_module_1 = require('./app.module');
+platform_browser_dynamic_1.platformBrowserDynamic().bootstrapModule(app_module_1.AppModule);
 //# sourceMappingURL=main.js.map